Move map composition into the `EnrichmentMap` class
