+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/enrichment/EnrichmentFinder.java
@@ -19,30 +19,16 @@
  */
 package io.spine.tools.compiler.enrichment;
 
-import com.google.common.base.Joiner;
-import com.google.common.collect.HashMultimap;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Multimap;
 import com.google.protobuf.DescriptorProtos.DescriptorProto;
-import com.google.protobuf.DescriptorProtos.FieldDescriptorProto;
 import com.google.protobuf.DescriptorProtos.FileDescriptorProto;
-import io.spine.option.TypeNameParser;
-import io.spine.type.TypeName;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.AbstractMap;
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static io.spine.option.OptionsProto.enrichment;
-import static io.spine.option.OptionsProto.enrichmentFor;
-import static io.spine.option.RawListParser.getValueSeparator;
-
 /**
- * Finds event enrichment Protobuf definitions.
+ * Finds event enrichment definitions for messages defined in a single proto file.
  *
  * @author Alexander Litus
  * @author Alex Tymchenko
@@ -52,9 +38,6 @@ public class EnrichmentFinder {
     static final String PROTO_TYPE_SEPARATOR = ".";
 
     private final FileDescriptorProto file;
-    private final String packagePrefix;
-    private final TypeNameParser eventTypeParser;
-    private final TypeNameParser enrichmentTypeParser;
 
     /**
      * Creates a new instance.
@@ -63,13 +46,10 @@ public class EnrichmentFinder {
      */
     EnrichmentFinder(FileDescriptorProto file) {
         this.file = file;
-        this.packagePrefix = file.getPackage() + PROTO_TYPE_SEPARATOR;
-        this.eventTypeParser = new TypeNameParser(enrichmentFor, packagePrefix);
-        this.enrichmentTypeParser = new TypeNameParser(enrichment, packagePrefix);
     }
 
     /**
-     * Finds event enrichment Protobuf definitions in the file.
+     * Finds event enrichment definitions in the file.
      *
      * @return a map from enrichment type name to event to enrich type name
      */
@@ -78,121 +58,12 @@ public class EnrichmentFinder {
         log.debug("Looking up for the enrichments in {}", file.getName());
 
         final List<DescriptorProto> messages = file.getMessageTypeList();
-        final Map<String, String> result = findEnrichments(messages);
+        final String packagePrefix = file.getPackage() + PROTO_TYPE_SEPARATOR;
+        final EnrichmentMap map = new EnrichmentMap(packagePrefix);
+        final Map<String, String> result = map.findEnrichments(messages);
+
+        log.debug("Found {} enrichments", result.size());
         return result;
-    }
-
-    private Map<String, String> findEnrichments(List<DescriptorProto> messages) {
-        final HashMultimap<String, String> multimap = HashMultimap.create();
-        for (DescriptorProto msg : messages) {
-            putEntry(multimap, msg);
-        }
-        log().debug("Found enrichments: {}", multimap.toString());
-        final Map<String, String> merged = EnrichmentMap.mergeDuplicateValues(multimap);
-        return merged;
-    }
-
-    private static void put(Map.Entry<String, String> entry, Multimap<String, String> targetMap) {
-        // Put key and value separately to avoid an error.
-        targetMap.put(entry.getKey(), entry.getValue());
-    }
-
-    @SuppressWarnings("MethodWithMultipleLoops")  // It's fine as we don't expect too many items.
-    private void putEntry(Multimap<String, String> targetMap, DescriptorProto msg) {
-        final Map<String, String> entries = scanMsg(msg);
-        for (Map.Entry<String, String> entry : entries.entrySet()) {
-            put(entry, targetMap);
-        }
-        if (!entries.isEmpty()) {
-            return;
-        }
-        final Map<String, String> entryFromField = scanFields(msg);
-        if (entryFromField.size() > 0) {
-            for (Map.Entry<String, String> entry : entryFromField.entrySet()) {
-                put(entry, targetMap);
-            }
-            return;
-        }
-        final Map.Entry<String, String> entryFromInnerMsg = scanInnerMessages(msg);
-        if (entryFromInnerMsg != null) {
-            put(entryFromInnerMsg, targetMap);
-            log().debug("Found enrichment: {} -> {}",
-                        entryFromInnerMsg.getKey(),
-                        entryFromInnerMsg.getValue());
-        } else {
-            log().debug("No enrichment or event annotations found for message {}", msg.getName());
-        }
-    }
-
-    @SuppressWarnings("MethodWithMoreThanThreeNegations")
-    private Map<String, String> scanMsg(DescriptorProto msg) {
-        final ImmutableMap.Builder<String, String> result = ImmutableMap.builder();
-        final String messageName = packagePrefix + msg.getName();
-
-        final Logger log = log();
-        // Treating current {@code msg} as an enrichment object.
-        log.debug("Scanning message {} for the enrichment annotations", messageName);
-        final Collection<TypeName> eventTypes = eventTypeParser.parseUnknownOption(msg);
-        if (!eventTypes.isEmpty()) {
-            final String mergedValue = Joiner.on(getValueSeparator())
-                                             .join(eventTypes);
-            log.debug("Found target events: {}", mergedValue);
-            result.put(messageName, mergedValue);
-        } else {
-            log.debug("No target events found");
-        }
-
-        // Treating current {@code msg} as a target for enrichment (e.g. Spine event).
-        log.debug("Scanning message {} for the enrichment target annotations", messageName);
-        final Collection<TypeName> enrichmentTypes = enrichmentTypeParser.parseUnknownOption(msg);
-        if (!enrichmentTypes.isEmpty()) {
-            log.debug("Found enrichments for event {}: {}", messageName, enrichmentTypes);
-            for (TypeName enrichmentType : enrichmentTypes) {
-                final String typeNameValue = enrichmentType.value();
-                result.put(typeNameValue, messageName);
-            }
-        } else {
-            log.debug("No enrichments for event {} found", messageName);
-        }
-
-        return result.build();
-    }
-
-    private Map<String, String> scanFields(DescriptorProto msg) {
-        final String msgName = msg.getName();
-        log().debug("Scanning fields of message {} for the enrichment annotations", msgName);
-        final Map<String, String> enrichmentsMap = new HashMap<>();
-        for (FieldDescriptorProto field : msg.getFieldList()) {
-            if (ByOption.isSetFor(field)) {
-                final ByOption by = new ByOption(packagePrefix, msg, field);
-                final Map.Entry<String, String> foundEvents = by.collect();
-                enrichmentsMap.put(foundEvents.getKey(), foundEvents.getValue());
-            }
-        }
-        return enrichmentsMap;
-    }
-
-
-    @SuppressWarnings("MethodWithMultipleLoops")    // It's fine in this case.
-    private Map.Entry<String, String> scanInnerMessages(DescriptorProto msg) {
-        final Logger log = log();
-        log.debug("Scanning inner messages of {} message for the annotations", msg.getName());
-        for (DescriptorProto innerMsg : msg.getNestedTypeList()) {
-            for (FieldDescriptorProto field : innerMsg.getFieldList()) {
-                if (ByOption.isSetFor(field)) {
-                    final String outerEventName = packagePrefix + msg.getName();
-                    final String enrichmentName =
-                            outerEventName +
-                                    PROTO_TYPE_SEPARATOR +
-                                    innerMsg.getName();
-                    log.debug("'by' option found on field {} targeting outer event {}",
-                                field.getName(),
-                                outerEventName);
-                    return new AbstractMap.SimpleEntry<>(enrichmentName, outerEventName);
-                }
-            }
-        }
-        return null;
     }
 
     private static Logger log() {
