+++ gradle-plugins/model-compiler/src/main/java/io/spine/tools/compiler/enrichment/EnrichmentMap.java
@@ -23,26 +23,60 @@ package io.spine.tools.compiler.enrichment;
 import com.google.common.base.Joiner;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Multimap;
+import com.google.protobuf.DescriptorProtos;
+import com.google.protobuf.DescriptorProtos.DescriptorProto;
+import io.spine.option.TypeNameParser;
+import io.spine.type.TypeName;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.AbstractMap;
+import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
+import static io.spine.option.OptionsProto.enrichment;
+import static io.spine.option.OptionsProto.enrichmentFor;
 import static io.spine.option.RawListParser.getValueSeparator;
+import static io.spine.tools.compiler.enrichment.EnrichmentFinder.PROTO_TYPE_SEPARATOR;
 
 /**
+ * Composes enrichment map for multiple message declarations.
+ *
+ * @author Alexander Litus
+ * @author Alex Tymchenko
  * @author Alexander Yevsyukov
  */
 class EnrichmentMap {
 
     private static final String EMPTY_TYPE_NAME = "";
 
-    private EnrichmentMap() {
+    private final String packagePrefix;
+    private final TypeNameParser eventTypeParser;
+    private final TypeNameParser enrichmentTypeParser;
+
+    EnrichmentMap(String packagePrefix) {
+        this.packagePrefix = packagePrefix;
+        this.eventTypeParser = new TypeNameParser(enrichmentFor, packagePrefix);
+        this.enrichmentTypeParser = new TypeNameParser(enrichment, packagePrefix);
+    }
+
+    Map<String, String> findEnrichments(Iterable<DescriptorProto> messages) {
+        final HashMultimap<String, String> multimap = HashMultimap.create();
+        for (DescriptorProto msg : messages) {
+            handleMessage(multimap, msg);
+        }
+        log().debug("Found enrichments: {}", multimap.toString());
+        final Map<String, String> merged = merge(multimap);
+        return merged;
     }
 
     /**
-     * Merge duplicate values into a single value for the same key.
+     * Transforms the passed multimap with possible several entries per key, into
+     * a map where several values from the passed multimap are joined into
+     * a single value.
      *
      * <p>The values are joined with the
      * {@linkplain io.spine.option.RawListParser#VALUE_SEPARATOR value separator}.
@@ -51,13 +85,13 @@ class EnrichmentMap {
      * i.e. when processing a single enrichment type as a map key, but multiple target
      * event types as values.
      */
-    static Map<String, String> mergeDuplicateValues(HashMultimap<String, String> source) {
-        log().debug("Merging duplicate properties in enrichments.proto");
+    private static Map<String, String> merge(HashMultimap<String, String> source) {
+        log().debug("Merging duplicating entries");
         final ImmutableMap.Builder<String, String> mergedResult = ImmutableMap.builder();
         for (String key : source.keySet()) {
             final Set<String> valuesPerKey = source.get(key);
-            // Empty type name might be present in the values
-            // If so, remove it from the set
+            /* Empty type name might be present in the values.
+               If so, remove it from the set */
             valuesPerKey.remove(EMPTY_TYPE_NAME);
 
             final String mergedValue;
@@ -72,6 +106,109 @@ class EnrichmentMap {
         }
 
         return mergedResult.build();
+    }
+
+    private static void put(Map.Entry<String, String> entry, Multimap<String, String> targetMap) {
+        // Put key and value separately to avoid an error.
+        targetMap.put(entry.getKey(), entry.getValue());
+    }
+
+    @SuppressWarnings("MethodWithMultipleLoops")  // It's fine as we don't expect too many items.
+    private void handleMessage(Multimap<String, String> targetMap, DescriptorProto msg) {
+        final Map<String, String> entries = scanMsg(msg);
+        for (Map.Entry<String, String> entry : entries.entrySet()) {
+            put(entry, targetMap);
+        }
+        if (!entries.isEmpty()) {
+            return;
+        }
+        final Map<String, String> entryFromField = scanFields(msg);
+        if (entryFromField.size() > 0) {
+            for (Map.Entry<String, String> entry : entryFromField.entrySet()) {
+                put(entry, targetMap);
+            }
+            return;
+        }
+        final Map.Entry<String, String> entryFromInnerMsg = scanInnerMessages(msg);
+        if (entryFromInnerMsg != null) {
+            put(entryFromInnerMsg, targetMap);
+            log().debug("Found enrichment: {} -> {}",
+                        entryFromInnerMsg.getKey(),
+                        entryFromInnerMsg.getValue());
+        } else {
+            log().debug("No enrichment or event annotations found for message {}", msg.getName());
+        }
+    }
+
+    @SuppressWarnings("MethodWithMoreThanThreeNegations")
+    private Map<String, String> scanMsg(DescriptorProto msg) {
+        final ImmutableMap.Builder<String, String> result = ImmutableMap.builder();
+        final String messageName = packagePrefix + msg.getName();
+
+        final Logger log = log();
+        // Treating current {@code msg} as an enrichment object.
+        log.debug("Scanning message {} for the enrichment annotations", messageName);
+        final Collection<TypeName> eventTypes = eventTypeParser.parseUnknownOption(msg);
+        if (!eventTypes.isEmpty()) {
+            final String mergedValue = Joiner.on(getValueSeparator())
+                                             .join(eventTypes);
+            log.debug("Found target events: {}", mergedValue);
+            result.put(messageName, mergedValue);
+        } else {
+            log.debug("No target events found");
+        }
+
+        // Treating current {@code msg} as a target for enrichment (e.g. Spine event).
+        log.debug("Scanning message {} for the enrichment target annotations", messageName);
+        final Collection<TypeName> enrichmentTypes = enrichmentTypeParser.parseUnknownOption(msg);
+        if (!enrichmentTypes.isEmpty()) {
+            log.debug("Found enrichments for event {}: {}", messageName, enrichmentTypes);
+            for (TypeName enrichmentType : enrichmentTypes) {
+                final String typeNameValue = enrichmentType.value();
+                result.put(typeNameValue, messageName);
+            }
+        } else {
+            log.debug("No enrichments for event {} found", messageName);
+        }
+
+        return result.build();
+    }
+
+    private Map<String, String> scanFields(DescriptorProto msg) {
+        final String msgName = msg.getName();
+        log().debug("Scanning fields of message {} for the enrichment annotations", msgName);
+        final Map<String, String> enrichmentsMap = new HashMap<>();
+        for (DescriptorProtos.FieldDescriptorProto field : msg.getFieldList()) {
+            if (ByOption.isSetFor(field)) {
+                final ByOption by = new ByOption(packagePrefix, msg, field);
+                final Map.Entry<String, String> foundEvents = by.collect();
+                enrichmentsMap.put(foundEvents.getKey(), foundEvents.getValue());
+            }
+        }
+        return enrichmentsMap;
+    }
+
+
+    @SuppressWarnings("MethodWithMultipleLoops")    // It's fine in this case.
+    private Map.Entry<String, String> scanInnerMessages(DescriptorProto msg) {
+        final Logger log = log();
+        log.debug("Scanning inner messages of {} message for the annotations", msg.getName());
+        for (DescriptorProto innerMsg : msg.getNestedTypeList()) {
+            for (DescriptorProtos.FieldDescriptorProto field : innerMsg.getFieldList()) {
+                if (ByOption.isSetFor(field)) {
+                    final String outerEventName = packagePrefix + msg.getName();
+                    final String enrichmentName =
+                            outerEventName +
+                                    PROTO_TYPE_SEPARATOR +
+                                    innerMsg.getName();
+                    log.debug("'by' option found on field {} targeting outer event {}",
+                              field.getName(),
+                              outerEventName);
+                    return new AbstractMap.SimpleEntry<>(enrichmentName, outerEventName);
+                }
+            }
+        }
+        return null;
     }
 
     private enum LogSingleton {
